Extracting pre-existing tree-based PartitionIndex implementation to its own package
