+++ internal/pindex/doc.go
@@ -1,0 +1,2 @@
+// Package pindex contains implementations of PartitionIndex
+package pindex
