+++ internal/pindex/tree/doc.go
@@ -1,0 +1,2 @@
+// Package tree provides a tree-based PartitionIndex which is suitable for reduction
+package tree
