+++ internal/pindex/tree/partition_tree_iterator.go
@@ -1,4 +1,4 @@
-package dataframe
+package tree
 
 import (
 	"sync"
