+++ internal/pindex/tree/partition_tree_test.go
@@ -1,4 +1,4 @@
-package dataframe
+package tree
 
 import (
 	"math/rand"
